调整表结构；新增支付记录表

+++ sql/wallet.sql
@@ -3,7 +3,7 @@ create table wallet
     id      bigint unsigned auto_increment,
 
     user_id varchar(30) not null,
-    balance bigint default 0 comment '用户余额，单位： 分',
+    balance int default 0 comment '用户余额，单位： 分',
 
     primary key (id),
     unique key (user_id)
